+++ src/from_row.rs
@@ -1,4 +1,4 @@
-use chrono::{TimeZone, Utc};
+use apalis_sql::{DateTime, DateTimeExt, TaskRow};
 
 #[derive(Debug)]
 pub(crate) struct SqliteTaskRow {
@@ -17,11 +17,11 @@ pub(crate) struct SqliteTaskRow {
     pub(crate) metadata: Option<String>,
 }
 
-impl TryInto<apalis_sql::from_row::TaskRow> for SqliteTaskRow {
+impl TryInto<TaskRow> for SqliteTaskRow {
     type Error = sqlx::Error;
 
-    fn try_into(self) -> Result<apalis_sql::from_row::TaskRow, Self::Error> {
-        Ok(apalis_sql::from_row::TaskRow {
+    fn try_into(self) -> Result<TaskRow, Self::Error> {
+        Ok(TaskRow {
             job: self.job,
             id: self
                 .id
@@ -37,28 +37,19 @@ impl TryInto<apalis_sql::from_row::TaskRow> for SqliteTaskRow {
                 .ok_or_else(|| sqlx::Error::Protocol("Missing attempts".into()))?
                 as usize,
             max_attempts: self.max_attempts.map(|v| v as usize),
-            run_at: self.run_at.map(|ts| {
-                Utc.timestamp_opt(ts, 0)
-                    .single()
-                    .ok_or_else(|| sqlx::Error::Protocol("Invalid run_at timestamp".into()))
-                    .unwrap()
-            }),
+            run_at: self
+                .run_at
+                .map(<DateTime as DateTimeExt>::from_unix_timestamp),
             last_result: self
                 .last_result
                 .map(|res| serde_json::from_str(&res).unwrap_or(serde_json::Value::Null)),
-            lock_at: self.lock_at.map(|ts| {
-                Utc.timestamp_opt(ts, 0)
-                    .single()
-                    .ok_or_else(|| sqlx::Error::Protocol("Invalid run_at timestamp".into()))
-                    .unwrap()
-            }),
+            lock_at: self
+                .lock_at
+                .map(<DateTime as DateTimeExt>::from_unix_timestamp),
             lock_by: self.lock_by,
-            done_at: self.done_at.map(|ts| {
-                Utc.timestamp_opt(ts, 0)
-                    .single()
-                    .ok_or_else(|| sqlx::Error::Protocol("Invalid run_at timestamp".into()))
-                    .unwrap()
-            }),
+            done_at: self
+                .done_at
+                .map(<DateTime as DateTimeExt>::from_unix_timestamp),
             priority: self.priority.map(|v| v as usize),
             metadata: self
                 .metadata